+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * Added History panel to iceleaf GUI
 * Added SubscribeAddressUpdates gRPC api to node to monitor addresses
+* Added MonitorTool to use SubscribeAddressUpdates to monitor addresses
 
 ## 1.7.0